Generic Data
Changed the map and all classes for the map to utilise the stops classes and only fill from the list provided.

Map will display all stops assigned from the Controller assuming that its been assigned too correctly from the PrepareForSegue function of the previous Controller. If not assigned it will display an error Pin.

Only requires a view with the MapViewController assigned and  a valid MapView within.